+++ src/models/ParcelaNegociacao.ts
@@ -33,7 +33,7 @@ const ParcelaNegociacao = sequelize.define(
     },
     parcela: Sequelize.INTEGER,
     vencimento: Sequelize.DATE,
-    valorParcela: Sequelize.DECIMAL,
+    valorParcela: Sequelize.DOUBLE,
     dataPagamento: Sequelize.DATE,
     situacao: Sequelize.STRING,
     notificacao: Sequelize.JSONB,
